+++ stlearn/spatial.py
@@ -1,9 +1,0 @@
-from .spatials import SME, clustering, morphology, smooth, trajectory
-
-__all__ = [
-    "clustering",
-    "smooth",
-    "trajectory",
-    "morphology",
-    "SME",
-]