+++ stlearn/spatial/__init__.py
@@ -1,0 +1,15 @@
+# stlearn/spatial/__init__.py
+
+from . import SME
+from . import clustering
+from . import morphology
+from . import smooth
+from . import trajectory
+
+__all__ = [
+    "clustering",
+    "smooth",
+    "trajectory",
+    "morphology",
+    "SME",
+]