+++ stlearn/tl/label/__init__.py
@@ -1,0 +1,7 @@
+from .label import run_singleR, run_rctd, run_label_transfer
+
+__all__ = [
+    "run_singleR",
+    "run_rctd",
+    "run_label_transfer",
+]