+++ HISTORY.rst
@@ -16,6 +16,7 @@ API and Bug Fixes:
 * pl.cluster_plot - Does not keep colours from previous runs when clustering.
 * pl.trajectory.pseudotime_plot - Fix typing of cluster values in .uns["split_node"].
 * Removed datasets.example_bcba - Replaced with wrapper for scanpy.visium_sge.
+* Moved spatials directory to spatial.
 
 0.4.11 (2022-11-25)
 ------------------